Refactor thread list to fix title setting bug
The issue was that after integratin the thread list with the thread
search (i.e. when no search result, just 'search' by matching anything
and then sorting by date) the original thread list was actually unused.
Unknonwingly unused because other parts of the codebase were
invalidating it, but it was not what actually got read.

This fix resolves that by moving the searched thread list into a store.
It also paves the way for limiting the number of results (limit/offset).
So far i've seen zero slowdowns returning every single thread, however,
my computer is fast and it's not memory efficient to hold all threads
there.